add: implement client for package users
