--- internal/client/writer.go
@@ -1,15 +1,0 @@
-package client
-
-import "os"
-
-func (c *client) writeSetting(set interface{}, f *os.File) error {
-	return nil
-}
-
-func (c *client) writeRaw(raw interface{}, f *os.File) error {
-	return nil
-}
-
-func (c *client) writeAnalyzed(txt interface{}, f *os.File) error {
-	return nil
-}
